Remove the link from category of service approval to profiles.
Issue #949: Give providers the ability to "edit and resubmit" applications
